pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "argus-ai"
version = "0.1.0"
description = "Production-grade LLM observability. G-ARVIS scoring for Groundedness, Accuracy, Reliability, Variance, Inference Cost, and Safety."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "Anil Prasad", email = "anil@ambharii.com" },
]
keywords = [
    "llm",
    "observability",
    "monitoring",
    "ai-safety",
    "mlops",
    "garvis",
    "evaluation",
    "production-ai",
    "guardrails",
    "drift-detection",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "numpy>=1.24",
    "structlog>=23.0",
]

[project.optional-dependencies]
opentelemetry = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]
prometheus = [
    "prometheus-client>=0.17",
]
anthropic = [
    "anthropic>=0.25",
]
openai = [
    "openai>=1.0",
]
all = [
    "argus-ai[opentelemetry,prometheus,anthropic,openai]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.0",
]

[project.urls]
Homepage = "https://github.com/anilatambharii/argus-ai"
Documentation = "https://argus-ai.ambharii.com"
Repository = "https://github.com/anilatambharii/argus-ai"
Issues = "https://github.com/anilatambharii/argus-ai/issues"
Changelog = "https://github.com/anilatambharii/argus-ai/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/argus_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
